+++ app/src/main/java/com/mysore/ashtanga/yoga/yogarepublic/SundayFragment.kt
@@ -62,7 +62,7 @@ class SundayFragment : Fragment() {
 
         timeTable!!.setShowHeader(true)
         timeTable!!.setTableMode(TimeTableView.TableMode.SHORT)
-        val begining = SharedDate.sundayBig[0].getValue("start") as Long
+        val begining = (SharedDate.sundayBig[0].getValue("start") as Double).toLong()
         val date = Date(begining)
         val format = SimpleDateFormat("yyyy.MM.dd")
         val dateBegining = format.format(date)
@@ -90,16 +90,13 @@ class SundayFragment : Fragment() {
         SharedDate.sundaySmall.forEach {
 
             val col = selectColor(it.getValue("name").toString())
-
-            val st = Date(it.getValue("start") as Long)
+            val st = Date((it.getValue("start") as Double).toLong())
             val startTime = SimpleDateFormat("HH:mm").format(st)
-            val et = Date(it.getValue("end") as Long)
+            val et = Date((it.getValue("end") as Double).toLong())
             val endTime = SimpleDateFormat("HH:mm").format(et)
-
-
             var td1 = TimeData(0, "$startTime - $endTime\n${it.getValue("name")} - ${it.getValue("teacher")}",
-                col, R.color.white, it.getValue("start") as Long,
-                it.getValue("end") as Long)
+                col, R.color.white, (it.getValue("start") as Double).toLong(),
+                (it.getValue("end") as Double).toLong())
             values.add(td1)
 
         }
@@ -118,18 +115,14 @@ class SundayFragment : Fragment() {
 
         SharedDate.sundayBig.forEach {
 
-
             val col = selectColor(it.getValue("name").toString())
-
-            val st = Date(it.getValue("start") as Long)
+            val st = Date((it.getValue("start") as Double).toLong())
             val startTime = SimpleDateFormat("HH:mm").format(st)
-            val et = Date(it.getValue("end") as Long)
+            val et = Date((it.getValue("end") as Double).toLong())
             val endTime = SimpleDateFormat("HH:mm").format(et)
-
-
             var td1 = TimeData(0, "$startTime - $endTime\n${it.getValue("name")} - ${it.getValue("teacher")}",
-                col, R.color.white, it.getValue("start") as Long,
-                it.getValue("end") as Long)
+                col, R.color.white, (it.getValue("start") as Double).toLong(),
+                (it.getValue("end") as Double).toLong())
             values2.add(td1)
 
         }
